refactor(repair-conflicts): rename exports and files to purge-conflicting-unmined
Pure rename commit, no logic change. Sets up the following commit which
replaces the classification machinery with a surgical purge of records
where Conflicting=true and UnminedSince>0.

- stores/utxo/repair_conflicts.go → purge_conflicting_unmined.go
- stores/utxo/tests/repair_conflicts_test.go → purge_conflicting_unmined_test.go
- cmd/repairconflicts/ → cmd/purgeconflictingunmined/
- RepairConflictingChains → PurgeConflictingUnmined
- RepairReport → PurgeReport
- RepairOptions → PurgeOptions
- RepairProgressFunc → PurgeProgressFunc
- cmd wrapper RepairConflicts → PurgeConflictingUnmined

The "repair-conflicts" CLI subcommand keeps its name here; a later commit
renames it to "purge-conflicting-unmined" along with the operator-facing
log strings.

errors.ErrRepairNeeded / NewRepairNeededError are intentionally retained
— the error semantic ("operator intervention required") is unchanged and
renaming would ripple through ~10 test files for no functional gain.
