+++ stores/utxo/tests/purge_conflicting_unmined_test.go
@@ -89,7 +89,7 @@ func TestRepairConflictingChains_CleanState(t *testing.T) {
 	store := setupSQLiteFileStore(ctx, t)
 	querier := newQuerier()
 
-	report, err := utxo.RepairConflictingChains(ctx, store, querier, false, nil)
+	report, err := utxo.PurgeConflictingUnmined(ctx, store, querier, false, nil)
 	require.NoError(t, err)
 	require.Equal(t, 0, report.CaseAFixed)
 	require.Equal(t, 0, report.CaseCFixed)
@@ -131,7 +131,7 @@ func TestRepairConflictingChains_CaseA_SingleLoser(t *testing.T) {
 	require.NoError(t, err)
 
 	querier := newQuerier()
-	report, err := utxo.RepairConflictingChains(ctx, store, querier, false, nil)
+	report, err := utxo.PurgeConflictingUnmined(ctx, store, querier, false, nil)
 	require.NoError(t, err)
 	// Non-fatal errors may occur when parentTx's input references an external tx not in the store.
 	// The repair should still detect and fix TX_B.
@@ -195,7 +195,7 @@ func TestRepairConflictingChains_CaseA_CascadeToChildren(t *testing.T) {
 	require.NoError(t, err)
 
 	querier := newQuerier()
-	report, err := utxo.RepairConflictingChains(ctx, store, querier, false, nil)
+	report, err := utxo.PurgeConflictingUnmined(ctx, store, querier, false, nil)
 	require.NoError(t, err)
 	// Non-fatal errors from external (unresolvable) parent inputs are expected; repair still runs.
 	require.Equal(t, 1, report.CaseAFixed, "only the root loser TX_B is the Case A entry")
@@ -239,7 +239,7 @@ func TestRepairConflictingChains_DryRun(t *testing.T) {
 	querier := newQuerier()
 
 	// Dry run — must report CaseA but not fix.
-	report, err := utxo.RepairConflictingChains(ctx, store, querier, true, nil)
+	report, err := utxo.PurgeConflictingUnmined(ctx, store, querier, true, nil)
 	require.NoError(t, err)
 	require.Equal(t, 1, report.CaseAFixed, "dry run must still report CaseAFixed")
 
@@ -280,7 +280,7 @@ func TestRepairConflictingChains_UnminedSinceFix(t *testing.T) {
 		blockHeaderIDs:  []uint32{1},
 	}
 
-	report, err := utxo.RepairConflictingChains(ctx, store, querier, false, nil)
+	report, err := utxo.PurgeConflictingUnmined(ctx, store, querier, false, nil)
 	require.NoError(t, err)
 	// SQL store returns nil from ScanInconsistentUnminedTxs — step 0 is skipped.
 	require.Equal(t, 0, report.UnminedSinceFixed)
@@ -359,7 +359,7 @@ func TestRepairConflictingChains_CaseC_InvertedWinnerLoser(t *testing.T) {
 		blockHeaderIDs:  []uint32{bestChainBlockID},
 	}
 
-	report, err := utxo.RepairConflictingChains(ctx, store, querier, false, nil)
+	report, err := utxo.PurgeConflictingUnmined(ctx, store, querier, false, nil)
 	require.NoError(t, err)
 	require.Equal(t, 1, report.CaseCFixed, "Case C should be detected and fixed")
 	require.Equal(t, 0, report.CaseAFixed, "TX_A's loser state is handled by ProcessConflicting in Case C, not Case A")
@@ -415,7 +415,7 @@ func TestRepairConflictingChains_CaseC_DryRun(t *testing.T) {
 	}
 
 	// dry-run: report but don't write
-	report, err := utxo.RepairConflictingChains(ctx, store, querier, true, nil)
+	report, err := utxo.PurgeConflictingUnmined(ctx, store, querier, true, nil)
 	require.NoError(t, err)
 	require.Equal(t, 1, report.CaseCFixed, "Case C should be reported")
 
@@ -480,7 +480,7 @@ func TestRepairConflictingChains_CaseASkippedAfterCaseC(t *testing.T) {
 		blockHeaderIDs:  []uint32{bestChainBlockID},
 	}
 
-	report, err := utxo.RepairConflictingChains(ctx, store, querier, false, nil)
+	report, err := utxo.PurgeConflictingUnmined(ctx, store, querier, false, nil)
 	require.NoError(t, err)
 	require.Equal(t, 1, report.CaseCFixed)
 	require.Equal(t, 0, report.CaseAFixed, "TX_A must not be double-counted: Case C ProcessConflicting already fixed it")
@@ -546,7 +546,7 @@ func TestRepairConflictingChains_CaseD_OrphanConflictingParent(t *testing.T) {
 		blockHeaderIDs:  []uint32{1},
 	}
 
-	report, err := utxo.RepairConflictingChains(ctx, store, querier, false, nil)
+	report, err := utxo.PurgeConflictingUnmined(ctx, store, querier, false, nil)
 	require.NoError(t, err)
 	require.Equal(t, 1, report.CaseDFixed, "parent orphan-conflicting must be detected and unmarked")
 	require.Equal(t, 0, report.CaseAFixed, "child is the recorded winner; no Case A")
@@ -602,7 +602,7 @@ func TestRepairConflictingChains_CaseD_DryRun(t *testing.T) {
 		blockHeaderIDs:  []uint32{1},
 	}
 
-	report, err := utxo.RepairConflictingChains(ctx, store, querier, true, nil)
+	report, err := utxo.PurgeConflictingUnmined(ctx, store, querier, true, nil)
 	require.NoError(t, err)
 	require.Equal(t, 1, report.CaseDFixed, "dryRun must still report Case D")
 
@@ -678,7 +678,7 @@ func TestRepairConflictingChains_CaseD_LegitConflictNotUnmarked(t *testing.T) {
 	require.Equal(t, *childOfLoser.TxIDChainHash(), *preCheck.SpendingDatas[0].TxID,
 		"precondition: parentLoser.SpendingDatas[0] must name childOfLoser")
 
-	report, err := utxo.RepairConflictingChains(ctx, store, querier, false, nil)
+	report, err := utxo.PurgeConflictingUnmined(ctx, store, querier, false, nil)
 	require.NoError(t, err)
 	require.Equal(t, 0, report.CaseDFixed, "legit conflicting parent must NOT be unmarked by Case D")
 	require.Equal(t, 1, report.CaseDCascaded, "legit conflicting parent with non-conflicting child must cascade")
@@ -767,7 +767,7 @@ func TestRepairConflictingChains_CaseD_ChainedOrphans(t *testing.T) {
 		blockHeaderIDs:  []uint32{1},
 	}
 
-	report, err := utxo.RepairConflictingChains(ctx, store, querier, false, nil)
+	report, err := utxo.PurgeConflictingUnmined(ctx, store, querier, false, nil)
 	require.NoError(t, err)
 	require.Equal(t, 2, report.CaseDFixed, "both parent and grandparent must be unmarked in one run")
 
@@ -858,7 +858,7 @@ func TestRepairConflictingChains_CaseD_LegitCascadeWithNilParentSD(t *testing.T)
 		blockHeaderIDs:  []uint32{1},
 	}
 
-	report, err := utxo.RepairConflictingChains(ctx, store, querier, false, nil)
+	report, err := utxo.PurgeConflictingUnmined(ctx, store, querier, false, nil)
 	require.NoError(t, err)
 	require.Equal(t, 0, report.CaseDFixed, "legit loser must NOT be unmarked")
 	require.Equal(t, 1, report.CaseDCascaded, "legit loser with nil SD must still cascade via tracked direct children")
@@ -942,7 +942,7 @@ func TestRepairConflictingChains_CaseC_StaleSiblingSkipped(t *testing.T) {
 		blockHeaderIDs:  []uint32{bestChainBlockID},
 	}
 
-	report, err := utxo.RepairConflictingChains(ctx, store, querier, false, nil)
+	report, err := utxo.PurgeConflictingUnmined(ctx, store, querier, false, nil)
 	require.NoError(t, err, "stale non-conflicting sibling must not cause a fatal error")
 	require.Equal(t, 0, report.CaseCFixed, "stale sibling must NOT be enqueued as Case C winner")
 }
@@ -1032,7 +1032,7 @@ func TestRepairConflictingChains_CaseD_OrphanBlocksParentDetection(t *testing.T)
 		blockHeaderIDs:  []uint32{1},
 	}
 
-	report, err := utxo.RepairConflictingChains(ctx, store, querier, false, nil)
+	report, err := utxo.PurgeConflictingUnmined(ctx, store, querier, false, nil)
 	require.NoError(t, err)
 	require.Equal(t, 2, report.CaseDFixed, "both parentX and blocker must be unmarked in one run")
 
@@ -1098,7 +1098,7 @@ func TestRepairConflictingChains_CaseD_CascadeDryRun(t *testing.T) {
 		blockHeaderIDs:  []uint32{1},
 	}
 
-	report, err := utxo.RepairConflictingChains(ctx, store, querier, true, nil)
+	report, err := utxo.PurgeConflictingUnmined(ctx, store, querier, true, nil)
 	require.NoError(t, err)
 	require.Equal(t, 1, report.CaseDCascaded, "dryRun must still report cascade")
